+++ src/core/proto/index.ts
@@ -1,3 +1,0 @@
-export * from './token-key'
-export * from './token-key-create'
-export * from './token-key-list-query'